refactor: extract KgsWidgetRenderer
